refactor: relocate ChangeTransportBoxState DTOs into use case subfolder
Move ChangeTransportBoxStateRequest and ChangeTransportBoxStateResponse
from Features/Logistics/UseCases/ into the ChangeTransportBoxState/
subfolder, matching the per-use-case layout used by every other use
case in the Logistics module. Update their namespaces to
...UseCases.ChangeTransportBoxState; remove the now-orphan
...UseCases using directives from the controller and three test files.

No logic, contract, MediatR, DI, or OpenAPI surface change — NSwag
identifies DTOs by class name, so the generated TypeScript client is
unchanged.
