+++ backend/src/Anela.Heblo.Application/Features/Logistics/UseCases/ChangeTransportBoxState/ChangeTransportBoxStateResponse.cs
@@ -1,7 +1,7 @@
 using Anela.Heblo.Application.Features.Logistics.UseCases.GetTransportBoxById;
 using Anela.Heblo.Application.Shared;
 
-namespace Anela.Heblo.Application.Features.Logistics.UseCases;
+namespace Anela.Heblo.Application.Features.Logistics.UseCases.ChangeTransportBoxState;
 
 public class ChangeTransportBoxStateResponse : BaseResponse
 {
