+++ backend/test/Anela.Heblo.Tests/Domain/Logistics/TransportBoxUniquenessTests.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 using Anela.Heblo.Application.Features.Catalog.Services;
-using Anela.Heblo.Application.Features.Logistics.UseCases;
 using Anela.Heblo.Application.Features.Logistics.UseCases.ChangeTransportBoxState;
 using Anela.Heblo.Application.Shared;
 using Anela.Heblo.Domain.Features.Catalog.Stock;
