+++ backend/test/Anela.Heblo.Tests/Features/Logistics/Transport/TransportBoxControllerTests.cs
@@ -1,5 +1,4 @@
 using Anela.Heblo.API.Controllers;
-using Anela.Heblo.Application.Features.Logistics.UseCases;
 using Anela.Heblo.Application.Features.Logistics.UseCases.ChangeTransportBoxState;
 using Anela.Heblo.Application.Features.Logistics.UseCases.OpenOrResumeBoxByCode;
 using Anela.Heblo.Application.Shared;
